==> jupyter_notebooks/04 -Hypothesis Testing.ipynb <==
# 04 – Hypothesis Testing & Statistical Evaluation

**Notebook Name:** `04_Hypothesis_Testing_Statistical_Evaluation.ipynb`

## Objectives
- Welch’s t-test: New vs Old builds.
- One-way ANOVA + Tukey: Price by property type.
- Optional t-test: London vs Rest of E&W.
- Report statistics, p-values, and decisions at α=0.05.

## Inputs
- `HousePricesRecords_clean.csv`

## Section 0: Libraries & Load Data

import pandas as pd
from scipy import stats
from statsmodels.stats.multicomp import pairwise_tukeyhsd
import matplotlib.pyplot as plt
import seaborn as sns


df_chunk = pd.read_csv("../outputs/datasets/collection/HousePricesRecords_clean.csv")
df_chunk.head()

## H1 - New vs Old Builds

new_prices = df_chunk[df_chunk["Old/New"] == 1]["Price"]
old_prices = df_chunk[df_chunk["Old/New"] == 0]["Price"]

t_stat, p_val = stats.ttest_ind(new_prices, old_prices, equal_var=False)

print(f"T-statistic: {t_stat:.2f}")
print(f"P-value: {p_val:.4f}")

if p_val < 0.05:
    print("Reject the null hypothesis – New houses are significantly more expensive.")
else:
    print("Fail to reject the null hypothesis – No significant difference in price.")

## H2 - Price by Property Type

property_types = ['Property_D', 'Property_F', 'Property_S', 'Property_T']
grouped_prices = [df_chunk[df_chunk[col] == 1]["Price"] for col in property_types if col in df_chunk.columns]


f_stat, p_val = stats.f_oneway(*grouped_prices)
print(f"F-statistic: {f_stat:.2f}")
print(f"P-value: {p_val:.4f}")

if p_val < 0.05:
    print("Reject the null hypothesis – Property type affects price.")
else:
    print("Fail to reject the null hypothesis – No significant difference between property types.")
